requirements.txt
ninja
peft
decord
einops
scikit-image
accelerate
deepspeed
prettytable
fastapi
torchvision
gradio==3.23
markdown2[all]
numpy
librosa
wandb
requests
sentencepiece
tokenizers==0.14
uvicorn
transformers==4.34.0
httpx==0.23.0
scikit-image
timm
iopath
omegaconf
decord
einops
ftfy
pytorchvideo
opencv-python
flash-attn
pycocotools
huggingface_hub